Safari-609.1.11.1/LayoutTests/js/dom/modules/module-not-found-error-event.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test script.onerror will be fired when the resource specified by the import declaration in the inlined module is not found.');
// Module will be executed asynchronously.
window.jsTestIsAsync = true;
debug('Module is not executed yet.');
var error = null;
function onError()
{
    error = window.event;
}
</script>
<script src="../../../resources/js-test-post.js"></script>
<script type="module" onerror="onError()">
import "./not-found.js";
testFailed('executed');
</script>
<script type="module">
shouldNotBe(`error`, `null`);
shouldBe(`error.type`, `"error"`);
finishJSTest()
</script>
</body>
</html>
